Add function that lets user change grid size
The function prompts the user to input a number (between 1-100),
it validates that the number is indeed in the expected range or else
it will keep looping prompts until a valid input is given.
Once a valid value is input, the current grid is removed and a new grid
is created with the amount of squares the user specified.
It's worth noting the squares resize themselves accordingly to fill out
the grid space