Change mouse event and add color/opacity function
Changed the mouse events from mouse*move* to mouse*enter*.
Added a function that generates a random RGB for a color to fill out
a square, as well as a function that adds 10 percent of opacity
each time the mouse enters a square's area (meaning if a specific
square has been "entered" 10 times, it will be fully colored).

Need to figure out how to make it so the random color generated is
permanent, meaning I won't get 10% opacity red, then 20% blue, then..
and rather have it stay on the original generated color and gradually
get darker.