Add color change for toggle and fix spam crashing
The "toggle opacity" button switches colors depending if it's OFF/ON.
Also, the grid changes only execute now if the input is indeed valid,
meaning there's not going to be attempts at creating in invalid grid
that caused the webpage to crash (such ass "1000" as input).
Moreover, made a functioning "current grid size" indicator which
changes dynamically according to the last user input (default is 16x16).